1、新增图像库映射

+++ utils-support-common-starter/src/main/java/com/chua/common/support/utils/FileUtils.java
@@ -251,6 +251,9 @@ public class FileUtils {
    * @throws IOException 删除异常
    */
   public static void forceDelete(final File file) throws IOException {
+    if(null == file) {
+      return;
+    }
     Objects.requireNonNull(file, "file");
     try {
       PathUtils.delete(file.toPath(), PathUtils.EMPTY_LINK_OPTION_ARRAY);
